JPA Repository Injeção de dependencia Database seeding

+++ src/main/java/com/example/training/jpazin/entities/User.java
@@ -14,7 +14,7 @@ public class User implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     private String name;
     private String email;
     private String phone;
@@ -25,7 +25,7 @@ public class User implements Serializable {
     public User() {
     }
 
-    public User(int id, String name, String email, String phone, String password) {
+    public User(Long id, String name, String email, String phone, String password) {
         this.id = id;
         this.name = name;
         this.email = email;
@@ -36,11 +36,11 @@ public class User implements Serializable {
 
 
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
